tipos de chains langchain/requirements.txt
langchain>=0.0.330
langchain-openai>=0.0.8
openai>=1.10.0
python-dotenv>=1.0.0  # Optional, for .env file support
pandas==2.2.0
langchain-ollama>=0.0.1